Managing flags

+++ j03/ex00/FragTrap.cpp
@@ -87,7 +87,7 @@ void FragTrap::takeDamage(unsigned int amount)
         std::cout << this->_name << " is already dead." << std::endl;
     else
     {
-        if (amount - this->_armorReduction > this->_maxHitPoints)
+        if (amount - this->_armorReduction > (unsigned int)this->_maxHitPoints)
             this->_energyPoints -= this->_maxHitPoints;
         else
             this->_energyPoints -= amount - this->_armorReduction;
@@ -105,7 +105,7 @@ void FragTrap::beRepaired(unsigned int amount)
         std::cout << this->_name << " is already full life." << std::endl;
     else
     {
-        if (amount + this->_energyPoints > this->_maxEnergyPoints)
+        if (amount + this->_energyPoints > (unsigned int)this->_maxEnergyPoints)
             this->_energyPoints = this->_maxEnergyPoints;
         else
             this->_energyPoints += amount;
