+++ j03/ex01/ScavTrap.cpp
@@ -88,7 +88,7 @@ void ScavTrap::takeDamage(unsigned int amount)
         std::cout << this->_name << " is already dead. (Again..)" << std::endl;
     else
     {
-        if (amount - this->_armorReduction > this->_maxHitPoints)
+        if (amount - this->_armorReduction > (unsigned int)this->_maxHitPoints)
             this->_energyPoints -= this->_maxHitPoints;
         else
             this->_energyPoints -= amount - this->_armorReduction;
@@ -106,7 +106,7 @@ void ScavTrap::beRepaired(unsigned int amount)
         std::cout << this->_name << " is already full life. (How can it be ?!)" << std::endl;
     else
     {
-        if (amount + this->_energyPoints > this->_maxEnergyPoints)
+        if (amount + this->_energyPoints > (unsigned int)this->_maxEnergyPoints)
             this->_energyPoints = this->_maxEnergyPoints;
         else
             this->_energyPoints += amount;
